backend/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==2.2.5
Flask-Cors==5.0.0
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.5
MarkupSafe==2.1.5
python-dotenv==0.21.1
requests==2.31.0
schedule==1.2.2
typing-extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
